requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
idna==3.10
psycopg2-binary==2.9.11
requests==2.32.5
tqdm==4.67.1
urllib3==2.5.0
